pyproject.toml
[project]
name = "dataextraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pyautogui>=0.9.54",
    "pydantic>=2.12.5",
    "pyinstaller>=6.17.0",
    "pyperclip>=1.11.0",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
]
